+++ bindings/ocaml/CHANGES.md
@@ -1,3 +1,10 @@
+## 0.4.0
+- Complete documentation
+- Redesigned API:
+  * by default, functions return result buffer instead of taking them as arguments
+  * arguments are now labeled
+- Improvements to runtime checks, unit tests
+
 ## 0.3.2
 - Build fixes addressing performance and compatibility issues